HGNN-Recommender-1/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.11
aiosignal==1.3.1
async-timeout==5.0.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.10.0
idna==3.10
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.1.0
networkx==3.1
numpy==1.24.4
propcache==0.2.0
psutil==6.1.0
pyparsing==3.1.4
PyYAML==6.0.2
requests==2.32.3
scipy==1.10.1
sympy==1.13.3
torch==2.4.1
torch-geometric==2.6.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.15.2
